N9 Исрпавил метод удаления товара чтобы соответствовало бизнес-правилам

+++ src/main/java/com/example/OrderManagment/Service/ProductServiceImpl.java
@@ -2,6 +2,9 @@ package com.example.OrderManagment.Service;
 
 import com.example.OrderManagment.Entity.Product;
 import com.example.OrderManagment.Entity.ProductStatus;
+import com.example.OrderManagment.Exception.BusinessException;
+import com.example.OrderManagment.Exception.ProductNotFoundException;
+import com.example.OrderManagment.Repository.OrderRepository;
 import com.example.OrderManagment.Repository.ProductRepository;
 import com.example.OrderManagment.dto.CreateProductRequestDto;
 import com.example.OrderManagment.dto.ProductResponseDto;
@@ -17,16 +20,20 @@ import java.util.Optional;
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
+    private final OrderRepository orderRepository;
     private final ProductMapper productMapper;
 
-    public ProductServiceImpl(ProductRepository productRepository, ProductMapper productMapper) {
+    public ProductServiceImpl(ProductRepository productRepository, OrderRepository orderRepository, ProductMapper productMapper) {
         this.productRepository = productRepository;
+        this.orderRepository = orderRepository;
         this.productMapper = productMapper;
     }
 
 
     @Override
     public ProductResponseDto createProduct(CreateProductRequestDto productRequestDto) {
+        // Тут можно добавить чтобы
+        // только менеджер мог создавать товар
         Product product = productMapper.toEntity(productRequestDto);
         Product product1 = productRepository.save(product);
         return productMapper.toDto(product1);
@@ -63,8 +70,15 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProduct(Long id) { // Тут добавлю реализацию чтобы соответствовало бизнес правилам
+    public void deleteProduct(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("Card not found: " + id);
+        }
+        if (orderRepository.existsByOrderItem_Product_Id(id)) {
+            throw new BusinessException("You cannot delete the product contained in the order");
+        }
         productRepository.deleteById(id);
+
     }
 
     @Override
